// ui/src/api/ora/common/v1/label_pb.ts
// @generated by protoc-gen-es v2.15.0 with parameter "target=ts"
// @generated from file ora/common/v1/label.proto (package ora.common.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file ora/common/v1/label.proto.
 */
export const file_ora_common_v1_label: GenFile = /*@__PURE__*/
  fileDesc("ChlvcmEvY29tbW9uL3YxL2xhYmVsLnByb3RvEg1vcmEuY29tbW9uLnYxIiMKBUxhYmVsEgsKA2tleRgBIAEoCRINCgV2YWx1ZRgCIAEoCSI4CgtMYWJlbEZpbHRlchILCgNrZXkYASABKAkSEgoFdmFsdWUYAiABKAlIAIgBAUIICgZfdmFsdWViBnByb3RvMw");

/**
 * A label for categorizing and grouping entities.
 *
 * @generated from message ora.common.v1.Label
 */
export type Label = Message<"ora.common.v1.Label"> & {
  /**
   * The key of the label.
   *
   * @generated from field: string key = 1;
   */
  key: string;

  /**
   * The value of the label.
   *
   * @generated from field: string value = 2;
   */
  value: string;
};

/**
 * Describes the message ora.common.v1.Label.
 * Use `create(LabelSchema)` to create a new message.
 */
export const LabelSchema: GenMessage<Label> = /*@__PURE__*/
  messageDesc(file_ora_common_v1_label, 0);

/**
 * A filter for labels.
 *
 * @generated from message ora.common.v1.LabelFilter
 */
export type LabelFilter = Message<"ora.common.v1.LabelFilter"> & {
  /**
   * The key of the label to filter by.
   *
   * @generated from field: string key = 1;
   */
  key: string;

  /**
   * An optional value of the label to filter by,
   * if not specified, entities with
   * the given key and any value will match.
   *
   * @generated from field: optional string value = 2;
   */
  value?: string | undefined;
};

/**
 * Describes the message ora.common.v1.LabelFilter.
 * Use `create(LabelFilterSchema)` to create a new message.
 */
export const LabelFilterSchema: GenMessage<LabelFilter> = /*@__PURE__*/
  messageDesc(file_ora_common_v1_label, 1);