// ui/src/api/ora/common/v1/time_range_pb.ts
// @generated by protoc-gen-es v2.15.0 with parameter "target=ts"
// @generated from file ora/common/v1/time_range.proto (package ora.common.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Timestamp } from "@bufbuild/protobuf/wkt";
import { file_google_protobuf_timestamp } from "@bufbuild/protobuf/wkt";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file ora/common/v1/time_range.proto.
 */
export const file_ora_common_v1_time_range: GenFile = /*@__PURE__*/
  fileDesc("Ch5vcmEvY29tbW9uL3YxL3RpbWVfcmFuZ2UucHJvdG8SDW9yYS5jb21tb24udjEiXwoJVGltZVJhbmdlEikKBXN0YXJ0GAEgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBInCgNlbmQYAiABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wYgZwcm90bzM", [file_google_protobuf_timestamp]);

/**
 * A time range.
 *
 * A time range is a range of time between two points.
 *
 * The time range is inclusive of the start time and exclusive of the end time.
 *
 * @generated from message ora.common.v1.TimeRange
 */
export type TimeRange = Message<"ora.common.v1.TimeRange"> & {
  /**
   * The start time of the time range.
   *
   * If not provided, the time range is open-ended.
   *
   * @generated from field: google.protobuf.Timestamp start = 1;
   */
  start?: Timestamp | undefined;

  /**
   * The end time of the time range.
   *
   * If not provided, the time range is open-ended.
   *
   * @generated from field: google.protobuf.Timestamp end = 2;
   */
  end?: Timestamp | undefined;
};

/**
 * Describes the message ora.common.v1.TimeRange.
 * Use `create(TimeRangeSchema)` to create a new message.
 */
export const TimeRangeSchema: GenMessage<TimeRange> = /*@__PURE__*/
  messageDesc(file_ora_common_v1_time_range, 0);